utils.js moved to utils folder
