+++ js/screens/game-1.js
@@ -1,7 +1,7 @@
 import {
   makeElement,
   showScreen,
-} from "../utils";
+} from "../utils/utils";
 import headerTemplate from '../templates/header-template';
 import statsTemplate from '../templates/stats-template';
 import greeting from "./greeting";
